refactor(connections): extract establishment journeys
