+++ docs/context/architecture/auth-secrets.md
@@ -7,6 +7,8 @@ sources:
   - src/treg/oauth.py
   - src/treg/oauth_providers.py
   - src/treg/health.py
+  - src/treg/application/connect.py
+  - src/treg/routers/connections.py
   - src/treg/routers/resources.py
 related:
   - architecture/proxy-model.md
@@ -66,7 +68,9 @@ hit it live). Now the method rides in the blob and `refresh()` honors it; a lega
 field gets body auth, then ONE retry with Basic on a 4xx, and stamps whichever worked — existing
 broken connections self-heal on their next call, no migration.
 
-**Connect flow (mint the first token):** `consent_url(pending)` builds the provider consent URL
+**Connect flow (mint the first token):** `application.connect` owns connection-establishment sessions
+and commit ordering, while `routers.connections` translates HTTP input and framework-neutral outcomes.
+`consent_url(pending)` builds the provider consent URL
 (default `access_type=offline` + `prompt=consent` so a refresh token comes back); `exchange_code(pending,
 code, client)` trades the auth code for tokens and returns a self-refreshable blob. Both honor
 per-provider quirks carried on the `PendingOAuth` (snapshotted from the registry entry, below): a provider's
